Create readme.md
Romey Coles
IT1025 D-fall19-CRN 82800
Professor Sebold

                                                                                        Unit 2 

 
Hardware 

Latency is the average time for the sector being accessed to rotate into position under a head, after a completed seek.
The disk transfer rate (sometimes called media rate) is the speed at which data is transferred to and from the disk media (actual disk platter) and is a function of the recording frequency. 
It is generally described in megabytes per second (MBps). Modern hard disks have an increasing range of disk transfer rates from the inner diameter to the outer diameter of the disk.
Its difference is that the transfer rate moves data and the latency seeks the data. 

How a solid state (SSD) hard drive is different from a traditional hard drive. 
Solid state drives address many of the timing and structural problems that a hard drive has.
It functions as a hard drive yet is not mechanically made up the same way, yet still able to accomplish the same storing function with less difficulty and faster.

Ram & Ram access memory
Ram stands for random access memory. It is placed on a motherboard.
It is the ram that enables a program to run. Ram amounts determine the speed a computer runs.
 
i.	Why does increasing RAM make the computer run faster? 
It Cause it keeps it from having to go back to the slower hard drive to transfer the data.
 
ii.	What is the difference between 64 and 32 bit data paths?
       It Refers to the number of bits of data transferred at a time or 1 clock cycle.
                     Dims have  64 bit data path it transferred 64 bits of data at a time.